+++ swap/src/protocol/alice/swap.rs
@@ -1,5 +1,7 @@
 //! Run an XMR/BTC swap in the role of Alice.
 //! Alice holds XMR and wishes receive BTC.
+use std::time::Duration;
+
 use crate::asb::{EventLoopHandle, LatestRate};
 use crate::bitcoin::ExpiredTimelocks;
 use crate::env::Config;
@@ -111,23 +113,72 @@ where
             }
         }
         AliceState::BtcLocked { state3 } => {
-            match state3.expired_timelocks(bitcoin_wallet).await? {
-                ExpiredTimelocks::None { .. } => {
-                    // Record the current monero wallet block height so we don't have to scan from
-                    // block 0 for scenarios where we create a refund wallet.
-                    let monero_wallet_restore_blockheight = monero_wallet.block_height().await?;
+            // We will retry indefinitely to lock the Monero funds, until the swap is cancelled
+            // Sometimes locking the Monero can fail e.g due to the daemon not being fully synced
+            let backoff = backoff::ExponentialBackoffBuilder::new()
+                .with_max_elapsed_time(None)
+                .with_max_interval(Duration::from_secs(60))
+                .build();
 
-                    let transfer_proof = monero_wallet
-                        .transfer(state3.lock_xmr_transfer_request())
-                        .await?;
+            let transfer_proof = backoff::future::retry(backoff, || async {
+                // We check the status of the Bitcoin lock transaction
+                // If the swap is cancelled, there is no need to lock the Monero funds anymore
+                // because there is no way for the swap to succeed.
+                if !matches!(
+                    state3.expired_timelocks(bitcoin_wallet).await?,
+                    ExpiredTimelocks::None { .. }
+                ) {
+                    return Ok(None);
+                }
 
+                // Record the current monero wallet block height so we don't have to scan from
+                // block 0 for scenarios where we create a refund wallet.
+                let monero_wallet_restore_blockheight = monero_wallet
+                    .block_height()
+                    .await
+                    .inspect_err(|e| {
+                        tracing::warn!(
+                            swap_id = %swap_id,
+                            error = ?e,
+                            "Failed to get Monero wallet block height while trying to lock XMR. We will retry."
+                        )
+                    })
+                    .map_err(backoff::Error::transient)?;
+
+                // Lock the Monero
+                monero_wallet
+                    .transfer(state3.lock_xmr_transfer_request())
+                    .await
+                    .map(|proof| Some((monero_wallet_restore_blockheight, proof)))
+                    .inspect_err(|e| {
+                        tracing::warn!(
+                            swap_id = %swap_id,
+                            error = ?e,
+                            "Failed to lock Monero. Make sure your monero-wallet-rpc is connected to a synced daemon and enough funds are available. We will retry."
+                        )
+                    })
+                    .map_err(backoff::Error::transient)
+            })
+            .await?;
+
+            match transfer_proof {
+                // If the transfer was successful, we transition to the next state
+                Some((monero_wallet_restore_blockheight, transfer_proof)) => {
                     AliceState::XmrLockTransactionSent {
                         monero_wallet_restore_blockheight,
                         transfer_proof,
                         state3,
                     }
                 }
-                _ => AliceState::SafelyAborted,
+                // If we were not able to lock the Monero funds before the timelock expired,
+                // we can safely abort the swap because we did not lock any funds
+                None => {
+                    tracing::info!(
+                        swap_id = %swap_id,
+                        "We did not manage to lock the Monero funds before the timelock expired. Aborting swap."
+                    );
+                    AliceState::SafelyAborted
+                }
             }
         }
         AliceState::XmrLockTransactionSent {
@@ -175,6 +226,11 @@ where
                        state3,
                    }
                 },
+                // TODO: We should already listen for the encrypted signature here.
+                //
+                // If we send Bob the transfer proof, but for whatever reason we do not receive an acknoledgement from him
+                // we would be stuck in this state forever (deadlock). By listening for the encrypted signature here we
+                // can still proceed to the next state even if Bob does not respond with an acknoledgement.
                 result = tx_lock_status.wait_until_confirmed_with(state3.cancel_timelock) => {
                     result?;
                     AliceState::CancelTimelockExpired {
@@ -194,7 +250,6 @@ where
 
             select! {
                 biased; // make sure the cancel timelock expiry future is polled first
-
                 result = tx_lock_status.wait_until_confirmed_with(state3.cancel_timelock) => {
                     result?;
                     AliceState::CancelTimelockExpired {
@@ -224,6 +279,7 @@ where
             ExpiredTimelocks::None { .. } => {
                 let tx_lock_status = bitcoin_wallet.subscribe_to(state3.tx_lock.clone()).await;
                 match state3.signed_redeem_transaction(*encrypted_signature) {
+                    // TODO: We should retry publishing the redeem transaction if it fails
                     Ok(tx) => match bitcoin_wallet.broadcast(tx, "redeem").await {
                         Ok((_, subscription)) => match subscription.wait_until_seen().await {
                             Ok(_) => AliceState::BtcRedeemTransactionPublished { state3 },
@@ -404,5 +460,17 @@ pub(crate) fn is_complete(state: &AliceState) -> bool {
             | AliceState::BtcRedeemed
             | AliceState::BtcPunished { .. }
             | AliceState::SafelyAborted
+    )
+}
+
+/// This function is used to check if Alice is in a state where it is clear that she has already received the encrypted signature from Bob.
+/// This allows us to acknowledge the encrypted signature multiple times
+/// If our acknowledgement does not reach Bob, he might send the encrypted signature again.
+pub(crate) fn has_already_processed_enc_sig(state: &AliceState) -> bool {
+    matches!(
+        state,
+        AliceState::EncSigLearned { .. }
+            | AliceState::BtcRedeemTransactionPublished { .. }
+            | AliceState::BtcRedeemed
     )
 }